requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
black==21.12b0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Django==4.0
django-multiselectfield==0.1.12
Flask==2.0.2
geographiclib==1.52
geopy==2.2.0
googlemaps==4.5.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
psycopg2==2.9.3
PyJWT==2.3.0
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
tornado==6.1
twilio==7.4.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
